requirements.txt.txt
streamlit
opencv-python
pillow
numpy
matplotlib
